requirements.txt
numpy<2
opencv-python==4.8.1.78
flask==3.0.0
flask-cors==4.0.0
youtube-transcript-api
yt-dlp
pytube
youtube-dl
google-generativeai
